move split_pdf to module
